Placeが保持するstar_aveのデータ型をintegerからfloatに変更

+++ db/migrate/20210517225647_add_column_to_places.rb
@@ -1,5 +1,5 @@
 class AddColumnToPlaces < ActiveRecord::Migration[6.1]
   def change
-    add_column :places, :star_ave, :integer
+    add_column :places, :star_ave, :float
   end
 end
